// src/interfaces/CancerVejiga.tsx
export interface BladderCancers {
  patients: Patient[];
}


export interface Patient {
  id:                   number;
  dni:                  string;
  edad:                 number;
  sexo:                 string;
  nhis:                 string;
  fechacir:             string;
  obesidad:             string;
  hta:                  string;
  dm:                   string;
  tabaco:               string;
  hereda:               string;
  expoprofesional:      string
  clinica:              string
  citologias:           string
  numtumores:           number;
  tamtumoral:           string;
  aspectotumoral:       string;
  estadotumoralclinico: string;
  acarsinoinsitu:       string;
  gradotumoral:         string;
  permiacionvascular:   string;
  carcinomaUrotelial:   string;
  fhistoatipicas:       string;
  primario:             string;
  recidivante:          string;
  instalacionprevia:    string;
  tac:                  string;
  rtu:                  string;
  progresiva:           string;
  recidiva:             string;
  exitus:               string;
  nrecidivas:           number;
  evoldesfavorable:     string;
  is_active: true
}


// Generated by https://quicktype.io
// Generated by https://quicktype.io

export interface History {
  created_at:           string;
  id:                   number;
  hisID:                number[];
  obesidad:             string;
  hta:                  string;
  dm:                   string;
  tabaco:               string;
  hereda:               string;
  expoprofesional:      string;
  clinica:              string;
  citologias:           string;
  numtumores:           number;
  tamtumoral:           string;
  aspectotumoral:       string;
  estadotumoralclinico: string;
  acarsinoinsitu:       string;
  gradotumoral:         string;
  permiacionvascular:   string;
  carcinomaUrotelial:   string;
  fhistoatipicas:       string;
  primario:             string;
  recidivante:          string;
  instalacionprevia:    string;
  tac:                  string;
  rtu:                  string;
  progresiva:           string;
  recidiva:             string;
  exitus:               string;
  nrecidivas:           string;
  evoldesfavorable:     string;
}



  